// backend/src/usuarios/dto/create-usuario.dto.ts
export class CreateUsuarioDto {
  nombres: string;
  apellidos: string;
  rut: string;
  direccion: string;
  edad: number;
  fecha_nacimiento: string;
  antiguedad: number;
  contacto: string;
  redes_sociales: string;
  discapacidad: boolean;
  tipo_discapacidad: string;
  inscrito_registro_discapacidad: boolean;
  parte_programas_municipales: boolean;
  descripcion_negocio: string;
  otro_ingreso: boolean;
  sector_economico: string;
  formalizado: boolean;
  autorizacion_sanitaria: boolean;
  productos: string;
  local_establecido: boolean;
  direccion_local: string;
  trabaja_solo: boolean;
  dificultad_fisica: boolean;
  dificultad_psicologica: boolean;
  pertenece_agrupacion: boolean;
}